[363] Fix(Spending) : Spending 컬럼 수정

See merge request s09-fintech-finance-sub2/S09P22A704!103

+++ back/src/main/java/com/richminime/global/common/codef/CodefWebClient.java
@@ -159,7 +159,7 @@ public class CodefWebClient {
                             .category(jsonObject1.getString("resMemberStoreType"))
                             .cost(Long.valueOf(jsonObject1.getString("resUsedAmount")))
                             .spentDate(date)
-                            .storeNo(Integer.valueOf(jsonObject1.getString("resMemberStoreNo")))
+                            .storeNo(Long.valueOf(jsonObject1.getString("resMemberStoreNo")))
                             .build();
                 })
                 .collect(Collectors.toList());
